Split flaky test
